requirements.txt
imbalanced_learn==0.13.0
imblearn==0.0
matplotlib==3.7.2
numpy==1.24.3
opencv_python==4.8.0.74
opencv_python_headless==4.10.0.84
Pillow==11.2.1
scikit_learn==1.7.0
tensorflow==2.13.0
tensorflow_intel==2.13.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
ultralytics==8.3.154
